feat(expansion): add initial support for wildcard expansion
- TODO: Remove the file as its tasks, including wildcard expansion, are now being implemented.
- expansion/expand_wildcard.c: Add initial implementation of `expand_wildcard` to handle wildcard expansion by splitting the input string into words and spaces.
- expansion/expansion.c: Integrate wildcard expansion by calling `expand_wildcard` at the end of the `expander` function pipeline.
- includes/expansion.h: Declare the `expand_wildcard` function to make it accessible across expansion modules.
- includes/stdfcts.h: Enforce const-correctness for string function prototypes like `ft_strcmp` and `ft_strjoin` for improved type safety.
- includes/utils.h: Remove `ft_strchr` and `ft_substr` declarations to centralize standard function prototypes in `stdfcts.h`.
- utils/skip_spaces.c: Update `skip_spaces` function signature and internal counters from `int` to `size_t` for correct handling of string indices.

+++ utils/skip_spaces.c
@@ -6,15 +6,15 @@
 /*   By: yjaafar <yjaafar@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 08:39:36 by yjaafar           #+#    #+#             */
-/*   Updated: 2025/07/01 08:39:38 by yjaafar          ###   ########.fr       */
+/*   Updated: 2025/08/06 01:06:56 by yjaafar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "utils.h"
 
-int	skip_spaces(char *str, int i)
+size_t	skip_spaces(char *str, size_t i)
 {
-	int	j;
+	size_t	j;
 
 	j = 0;
 	while (ft_isspace(str[i + j]))
